requirements.txt
langchain
langchain-openai>=0.1.0
python-dateutil>=2.9.0

langchain-community          # loaders modernos
unstructured[all-docs]       # extracción robusta
pymupdf                       # PDFs rápidos
pandas
python-multipart              # para subida de archivos en FastAPI
fastapi
uvicorn[standard]
pydantic
unstructured[ocr]==0.10.25

pytesseract>=0.3
pdf2image>=1.17
pillow>=10
pillow-heif>=0.13   # o 
pi-heif>=0.11